notifications:(tab) hide notifications and telemetry
