+++ src/preferences/advanced_tab.html
@@ -6,7 +6,6 @@
     <link rel="stylesheet" href="../../assets/css/preferences.css">
     <script>
       const featflag = require('./featflag.js');
-      const preferences = require('./preferences.js');
       document.addEventListener("DOMContentLoaded", function(event) {
         featflag.updateDialog();
       });
