+++ src/preferences/general_tab.html
@@ -7,7 +7,7 @@
     <script>
       const preferences = require('./preferences.js');
       document.addEventListener("DOMContentLoaded", function(event) {
-        preferences.updateContentDialog();
+        preferences.updateDialog();
       });
     </script>
   </head>
@@ -17,14 +17,18 @@
         <input type="checkbox" id="startup" onclick="preferences.toggleSetting(this);" /><label> Start Workstation App on system startup</label>
       </div>
       <div class="section">
-        <input type="checkbox" id="telemetry" onclick="preferences.toggleSetting(this);" />
-        <label>Share my usage data with Chef
-          - Commands (without arguments)
-          - Connection method (WinRM or SSH)
-          - Host/client operating system and version
-          <!--<a href="#">View more usage data details</a>-->
-        </label>
+        <input type="checkbox" id="software_updates" onclick="preferences.toggleSetting(this);" /><label> Automatically check for software updates</label>
       </div>
+      <!-- @afiune To be implemented with https://github.com/chef/chef-workstation-app/issues/158 -->
+      <!--<div class="section">-->
+        <!--<input type="checkbox" id="telemetry" onclick="preferences.toggleSetting(this);" />-->
+        <!--<label>Share my usage data with Chef-->
+          <!--- Commands (without arguments)-->
+          <!--- Connection method (WinRM or SSH)-->
+          <!--- Host/client operating system and version-->
+          <!--<a href="#">View more usage data details</a>-->
+        <!--</label>-->
+      <!--</div>-->
     </div>
   </body>
 </html>
