+++ src/preferences/notifications_tab.html
@@ -13,9 +13,7 @@
   </head>
   <body>
     <div class="content">
-      <div class="section">
-        <input type="checkbox" id="software_updates" onclick="notifications.toggleSetting(this);" /><label> Automatically check for software updates</label>
-      </div>
+      <!-- @afiune implement this functionality -->
     </div>
   </body>
 </html>
