+++ src/preferences/preferences_bar.html
@@ -16,10 +16,11 @@
           <div class="icon general"></div>
           <div class="tab-sub-title">General</div>
         </div>
-        <div class="tab" id="notifications" onclick="preferences.switchTab('notifications')">
-          <div class="icon notification"></div>
-          <div class="tab-sub-title">Notifications</div>
-        </div>
+        <!-- @afiune Left here as an example to add more tabs like the notifications tab -->
+        <!--<div class="tab" id="notifications" onclick="preferences.switchTab('notifications')">-->
+          <!--<div class="icon notification"></div>-->
+          <!--<div class="tab-sub-title">Notifications</div>-->
+        <!--</div>-->
         <div class="tab" id="advanced" onclick="preferences.switchTab('advanced')">
           <div class="icon advanced"></div>
           <div class="tab-sub-title">Advanced</div>
